Configuração navegação entre menus e implementação validação de remoção

+++ HoursControl/Application/Services/EmployeeService.cs
@@ -7,17 +7,21 @@ namespace HoursControl.Application.Services
     public class EmployeeService : IEmployeeService
     {
         private readonly IEmployeeRepository _employeeRepository;
+        private readonly IRecordRepository _recordRepository;
 
-        public EmployeeService(IEmployeeRepository employeeRepository)
+        public EmployeeService(
+            IEmployeeRepository employeeRepository, 
+            IRecordRepository recordRepository)
         {
             _employeeRepository = employeeRepository;
+            _recordRepository = recordRepository;
         }
 
-        public Employee SerchRegister(Guid id)
+        public Employee SearchRegister(Guid id)
         {
-            return _employeeRepository.SerchRegister(id);
+            return _employeeRepository.SearchRegister(id);
         }
-        public Employee SerchForCPF(string cpf)
+        public Employee SearchForCPF(string cpf)
         {
             return _employeeRepository.SearchForCPF(cpf);
         }
@@ -36,7 +40,15 @@ namespace HoursControl.Application.Services
 
         public void RemoveRegister(Guid id)
         {
-            _employeeRepository.RemoveRegister(id);
+            if(_recordRepository.SearchForEmployee(id) != null) 
+            {
+                Console.WriteLine("Não é possível remover o funcionário, pois há tarefas cadastradas nesse CPF");
+            }
+            else
+            {
+                _employeeRepository.RemoveRegister(id);
+            }
+            
         }
 
         public void UpdateRegister(Employee record)
